# Add a check for ASCII characters in to_upper and to_lower