--- a/library/core/src/unicode/unicode_data.rs
+++ b/library/core/src/unicode/unicode_data.rs
@@ -549,16 +549,24 @@ pub mod white_space {
 #[rustfmt::skip]
 pub mod conversions {
     pub fn to_lower(c: char) -> [char; 3] {
-        match bsearch_case_table(c, LOWERCASE_TABLE) {
-            None => [c, '\0', '\0'],
-            Some(index) => LOWERCASE_TABLE[index].1,
+        if c.is_ascii() {
+            [(c as u8).to_ascii_lowercase() as char, '\0', '\0']
+        } else {
+            match bsearch_case_table(c, LOWERCASE_TABLE) {
+                None => [c, '\0', '\0'],
+                Some(index) => LOWERCASE_TABLE[index].1,
+            }
         }
     }
 
     pub fn to_upper(c: char) -> [char; 3] {
-        match bsearch_case_table(c, UPPERCASE_TABLE) {
-            None => [c, '\0', '\0'],
-            Some(index) => UPPERCASE_TABLE[index].1,
+        if c.is_ascii() {
+            [(c as u8).to_ascii_uppercase() as char, '\0', '\0']
+        } else {
+            match bsearch_case_table(c, UPPERCASE_TABLE) {
+                None => [c, '\0', '\0'],
+                Some(index) => UPPERCASE_TABLE[index].1,
+            }
         }
     }