@nagisa 2021-01-24
This should be accompanied with some benchmark outputs.