@mcastorina 2021-01-25
I didn't see to_upper / to_lower being benchmarked anywhere, so I added two more benches for them. Below are the results before and after this PR's changes.

## Before
```
test char::methods::bench_ascii_to_lowercase  ... bench:     248,590 ns/iter (+/- 1,181)
test char::methods::bench_ascii_to_uppercase  ... bench:     250,286 ns/iter (+/- 513)
```

## After
```
test char::methods::bench_ascii_to_lowercase  ... bench:     152,079 ns/iter (+/- 84)
test char::methods::bench_ascii_to_uppercase  ... bench:     154,737 ns/iter (+/- 680)
```