@gilescope 2021-01-26
Just linking this up with the internals chat on this subject: https://internals.rust-lang.org/t/to-upper-speed/13896/12