@mcastorina 2021-02-03
# Additional Benchmarks

## Before
```
test char::methods::bench_ascii_mix_to_lowercase                ... bench:     248,212 ns/iter (+/- 235)
test char::methods::bench_ascii_mix_to_uppercase                ... bench:     251,416 ns/iter (+/- 135)
test char::methods::bench_ascii_to_lowercase                    ... bench:     247,025 ns/iter (+/- 204)
test char::methods::bench_ascii_to_uppercase                    ... bench:     246,656 ns/iter (+/- 108)
test char::methods::bench_non_ascii_to_lowercase                ... bench:     248,828 ns/iter (+/- 121)
test char::methods::bench_non_ascii_to_uppercase                ... bench:     253,036 ns/iter (+/- 1,798)
```

## After
```
test char::methods::bench_ascii_mix_to_lowercase                ... bench:     151,013 ns/iter (+/- 248)
test char::methods::bench_ascii_mix_to_uppercase                ... bench:     154,812 ns/iter (+/- 475)
test char::methods::bench_ascii_to_lowercase                    ... bench:      53,337 ns/iter (+/- 197)
test char::methods::bench_ascii_to_uppercase                    ... bench:      56,142 ns/iter (+/- 104)
test char::methods::bench_non_ascii_to_lowercase                ... bench:     247,925 ns/iter (+/- 295)
test char::methods::bench_non_ascii_to_uppercase                ... bench:     251,799 ns/iter (+/- 503)
```