@joshtriplett 2021-02-03
That's quite definitive! It's a performance win for strings containing ASCII, even if they're not entirely ASCII, and it appears entirely neutral for strings containing non-ASCII.

There are likely other potential performance wins to be had here, but this is an obvious win with no apparent downside.

Let's see if it affects any of the existing benchmarks.

@bors try @rust-timer queue