@pickfire 2021-02-07
Does this have an impact on `String::to_uppercase`?