@gilescope 2021-02-07
Yes - a positive one. At the moment the string to_uppercase calls the char fn for each char and though I think there's room for optimistion there too but that's a tail for another PR.